+++ NEWS.md
@@ -1,4 +1,14 @@
-csquares v0.0.3.005 (Release date: 2024-06-24)
+csquares v0.0.4 (Release date: 2024-07-05)
+-------------
+
+  * Added function to expand wildcards
+  * Added function to match wildcards
+  * Added vignette about wildcards
+  * Added multiple methods for base and tidyverse S3 generics
+  * Added self tests
+  * Several minor corrections
+
+csquares v0.0.3 (Release date: 2024-06-24)
 -------------
 
   * Initial release